ratatui_widgets

Module block

Elements related to the Block base widget.

This holds everything needed to display and configure a Block.

In its simplest form, a Block is a border around another widget. It can have a title and padding.

Structs

Block

A widget that renders borders, titles, and padding around other widgets.

Padding

Defines the padding for a Block.

Enums

TitlePosition

Defines the position of the title.

Traits

BlockExt

An extension trait for Block that provides some convenience methods.