ratatui_widgets::borders

Enum BorderType

pub enum BorderType {
    Plain,
    Rounded,
    Double,
    Thick,
    LightDoubleDashed,
    HeavyDoubleDashed,
    LightTripleDashed,
    HeavyTripleDashed,
    LightQuadrupleDashed,
    HeavyQuadrupleDashed,
    QuadrantInside,
    QuadrantOutside,
}

The type of border of a Block.

See the borders method of Block to configure its borders.

Variants

Plain

A plain, simple border.

This is the default

Example

┌───────┐
│       │
└───────┘

Rounded

A plain border with rounded corners.

Example

╭───────╮
│       │
╰───────╯

Double

A doubled border.

Note this uses one character that draws two lines.

Example

╔═══════╗
║       ║
╚═══════╝

Thick

A thick border.

Example

┏━━━━━━━┓
┃       ┃
┗━━━━━━━┛

LightDoubleDashed

A light double-dashed border.

┌╌╌╌╌╌╌╌┐
╎       ╎
└╌╌╌╌╌╌╌┘

HeavyDoubleDashed

A heavy double-dashed border.

┏╍╍╍╍╍╍╍┓
╏       ╏
┗╍╍╍╍╍╍╍┛

LightTripleDashed

A light triple-dashed border.

┌┄┄┄┄┄┄┄┐
┆       ┆
└┄┄┄┄┄┄┄┘

HeavyTripleDashed

A heavy triple-dashed border.

┏┅┅┅┅┅┅┅┓
┇       ┇
┗┅┅┅┅┅┅┅┛

LightQuadrupleDashed

A light quadruple-dashed border.

┌┈┈┈┈┈┈┈┐
┊       ┊
└┈┈┈┈┈┈┈┘

HeavyQuadrupleDashed

A heavy quadruple-dashed border.

┏┉┉┉┉┉┉┉┓
┋       ┋
┗┉┉┉┉┉┉┉┛

QuadrantInside

A border with a single line on the inside of a half block.

Example

▗▄▄▄▄▄▄▄▖
▐       ▌
▐       ▌
▝▀▀▀▀▀▀▀▘

QuadrantOutside

A border with a single line on the outside of a half block.

Example

▛▀▀▀▀▀▀▀▜
▌       ▐
▌       ▐
▙▄▄▄▄▄▄▄▟

Implementations

impl BorderType

pub const fn border_symbols<'a>(border_type: Self) -> Set<'a>

Convert this BorderType into the corresponding Set of border symbols.

pub const fn to_border_set<'a>(self) -> Set<'a>

Convert this BorderType into the corresponding Set of border symbols.