ratatui_widgets

Module canvas

A Canvas and a collection of Shapes.

The Canvas is a blank space on which you can draw anything manually or use one of the predefined Shapes.

The available shapes are:

Circle: A basic circle

Line: A line between two points

Map: A world map

Points: A scatter of points

Rectangle: A basic rectangle

You can also implement your own custom Shapes.

Structs

Canvas

The Canvas widget provides a means to draw shapes (Lines, Rectangles, Circles, etc.) on a grid.

Circle

A circle with a given center and radius and with a given color

Context

Holds the state of the Canvas when painting to it.

Label

Label to draw some text on the canvas

Line

A line from (x1, y1) to (x2, y2) with the given color

Map

A world map

Painter

Painter is an abstraction over the Context that allows to draw shapes on the grid.

Points

A group of points with a given color

Rectangle

A rectangle to draw on a Canvas

Enums

MapResolution

Defines how many points are going to be used to draw a Map.

Traits

Shape

Something that can be drawn on a Canvas.