ratatui_widgets::chart

Struct Dataset

pub struct Dataset<'a> { /* private fields */ }

A group of data points

This is the main element composing a Chart.

A dataset can be named. Only named datasets will be rendered in the legend.

After that, you can pass it data with Dataset::data. Data is an array of f64 tuples ((f64, f64)), the first element being X and the second Y. It’s also worth noting that, unlike the Rect, here the Y axis is bottom to top, as in math.

You can also customize the rendering by using Dataset::marker and Dataset::graph_type.

Example

This example draws a red line between two points.

use ratatui::style::Stylize;
use ratatui::symbols::Marker;
use ratatui::widgets::{Dataset, GraphType};

let dataset = Dataset::default()
    .name("dataset 1")
    .data(&[(1., 1.), (5., 5.)])
    .marker(Marker::Braille)
    .graph_type(GraphType::Line)
    .red();

Implementations

impl<'a> Dataset<'a>

pub fn name<S>(self, name: S) -> Self
where S: Into<Line<'a>>,

Sets the name of the dataset

The dataset’s name is used when displaying the chart legend. Datasets don’t require a name and can be created without specifying one. Once assigned, a name can’t be removed, only changed

The name can be styled (see Line for that), but the dataset’s style will always have precedence.

This is a fluent setter method which must be chained or used as it consumes self

pub const fn data(self, data: &'a [(f64, f64)]) -> Self

Sets the data points of this dataset

Points will then either be rendered as scattered points or with lines between them depending on Dataset::graph_type.

Data consist in an array of f64 tuples ((f64, f64)), the first element being X and the second Y. It’s also worth noting that, unlike the Rect, here the Y axis is bottom to top, as in math.

This is a fluent setter method which must be chained or used as it consumes self

pub const fn marker(self, marker: Marker) -> Self

Sets the kind of character to use to display this dataset

You can use dots (•), blocks (█), bars (▄), braille (⠓, ⣇, ⣿) or half-blocks (█, ▄, and ▀). See symbols::Marker for more details.

Note Marker::Braille requires a font that supports Unicode Braille Patterns.

This is a fluent setter method which must be chained or used as it consumes self

pub const fn graph_type(self, graph_type: GraphType) -> Self

Sets how the dataset should be drawn

Chart can draw scatter, line or bar charts. A scatter chart draws only the points in the dataset, a line char draws a line between each point, and a bar chart draws a line from the x axis to the point. See GraphType for more details

This is a fluent setter method which must be chained or used as it consumes self

pub fn style<S: Into<Style>>(self, style: S) -> Self

Sets the style of this dataset

The given style will be used to draw the legend and the data points. Currently the legend will use the entire style whereas the data points will only use the foreground.

style accepts any type that is convertible to Style (e.g. Style, Color, or your own type that implements Into<Style>).

This is a fluent setter method which must be chained or used as it consumes self

Example

Dataset also implements Stylize which mean you can style it like so

use ratatui::style::Stylize;
use ratatui::widgets::Dataset;

let dataset = Dataset::default().red();