ratatui_widgets::scrollbar

Struct ScrollbarState

pub struct ScrollbarState { /* private fields */ }

A struct representing the state of a Scrollbar widget.

Important

It’s essential to set the content_length field when using this struct. This field represents the total length of the scrollable content. The default value is zero which will result in the Scrollbar not rendering.

For example, in the following list, assume there are 4 bullet points:

the content_length is 4

the position is 0

the viewport_content_length is 2

┌───────────────┐
│1. this is a   █
│   single item █
│2. this is a   ║
│   second item ║
└───────────────┘

If you don’t have multi-line content, you can leave the viewport_content_length set to the default and it’ll use the track size as a viewport_content_length.

Implementations

impl ScrollbarState

pub const fn new(content_length: usize) -> Self

Constructs a new ScrollbarState with the specified content length.

content_length is the total number of element, that can be scrolled. See ScrollbarState for more details.

pub const fn position(self, position: usize) -> Self

Sets the scroll position of the scrollbar.

This represents the number of scrolled items.

This is a fluent setter method which must be chained or used as it consumes self

pub const fn content_length(self, content_length: usize) -> Self

Sets the length of the scrollable content.

This is the number of scrollable items. If items have a length of one, then this is the same as the number of scrollable cells.

This is a fluent setter method which must be chained or used as it consumes self

pub const fn viewport_content_length( self, viewport_content_length: usize, ) -> Self

Sets the items’ size.

This is a fluent setter method which must be chained or used as it consumes self

pub const fn prev(&mut self)

Decrements the scroll position by one, ensuring it doesn’t go below zero.

pub fn next(&mut self)

Increments the scroll position by one, ensuring it doesn’t exceed the length of the content.

pub const fn first(&mut self)

Sets the scroll position to the start of the scrollable content.

pub const fn last(&mut self)

Sets the scroll position to the end of the scrollable content.

pub fn scroll(&mut self, direction: ScrollDirection)

Changes the scroll position based on the provided ScrollDirection.

pub const fn get_position(&self) -> usize

Returns the current position within the scrollable content.