ratatui_widgets

Module table

The Table widget is used to display multiple rows and columns in a grid and allows selecting one or multiple cells.

Structs

Cell

A Cell contains the Text to be displayed in a Row of a Table.

Row

A single row of data to be displayed in a Table widget.

Table

A widget to display data in formatted columns.

TableState

State of a Table widget

Enums

HighlightSpacing

This option allows the user to configure the “highlight symbol” column width spacing