ratatui_widgets

Module chart

The Chart widget is used to plot one or more Dataset in a cartesian coordinate system.

Structs

Axis

An X or Y axis for the Chart widget

Chart

A widget to plot one or more Dataset in a cartesian coordinate system

Dataset

A group of data points

Enums

GraphType

Used to determine which style of graphing to use

LegendPosition

Allow users to specify the position of a legend in a Chart